Update Syllabus_Checker_For_GUI.py
updated readability functionality so it now reads the filename and takes the course number and assigns it to a specific range of accepted values for the readability test results. it now also outputs better suggestions.

also added error handling for files that dont follow "courseName_courseNumber" or "courseName.courseNumber" format.